backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# ML dependencies
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
pytorch-grad-cam>=1.4.8

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Validation and security
pydantic>=2.0.0

# Monitoring and logging
python-json-logger>=2.0.0

# Development dependencies (optional)
pytest>=7.4.0
httpx>=0.25.0  # For testing FastAPI
